+++ src/compounds/legacy-product-table/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.4.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.legacy-product-table@1.3.1...@uswitch/trustyle.legacy-product-table@1.4.0) (2021-07-07)
+
+
+### Features
+
+* add preapproved banner to product tables ([c253d0d](https://github.com/uswitch/trustyle/commit/c253d0d))
+
+
+
+
+
 ## [1.3.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.legacy-product-table@1.3.0...@uswitch/trustyle.legacy-product-table@1.3.1) (2021-07-01)
 
 **Note:** Version bump only for package @uswitch/trustyle.legacy-product-table